Move mock_app to test_support.py
